Create dummybrowser
This script is used as a custom default web browser. Instead of a random click potentially opening an actual web browser, the URL provided to the script is copied to clipboard by xsel and also stored in a log file for convenience.

Note: /dev/shm/tmp doesn't probably exist by default.